requirements.txt
aiofiles==24.1.0
aiosmtplib==4.0.1
alembic==1.14.1
aniso8601==10.0.1
Authlib==1.6.0
bcrypt==4.2.1
bidict==0.23.1
blinker==1.9.0
boto3==1.40.16
botocore==1.40.16
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==45.0.5
Deprecated==1.2.18
distlib==0.3.9
dnspython==2.7.0
email_validator==2.2.0
et_xmlfile==2.0.0
filelock==3.17.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
flask-cors==6.0.1
Flask-Limiter==3.10.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security-Too==5.6.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
git-filter-repo==2.47.0
greenlet==3.1.1
h11==0.16.0
h2==4.2.0
hpack==4.1.0
Hypercorn==0.17.3
hyperframe==6.1.0
icalendar==6.3.1
idna==3.10
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
limits==4.0.1
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==4.0.0
marshmallow-sqlalchemy==1.4.2
mdurl==0.1.2
numpy==2.3.0
openpyxl==3.1.5
ordered-set==4.1.0
packaging==24.2
pandas==2.3.0
passlib==1.7.4
pillow==11.1.0
platformdirs==4.3.6
priority==2.0.0
psycopg2-binary==2.9.10
pycparser==2.22
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-engineio==4.12.2
python-socketio==5.13.0
pytz==2025.2
Quart==0.20.0
quart-cors==0.8.0
rich==13.9.4
s3transfer==0.13.1
setuptools==75.8.0
simple-websocket==1.1.0
six==1.17.0
SQLAlchemy==2.0.38
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.5.0
virtualenv==20.29.1
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
WTForms==3.2.1
